Fix for mingw

+++ conanfile.py
@@ -1,11 +1,10 @@
 # -*- coding: utf-8 -*-
 
 from conans import ConanFile, AutoToolsBuildEnvironment, tools
-from conans.errors import ConanException, ConanExceptionInUserConanfileMethod
+from conans.errors import ConanExceptionInUserConanfileMethod
 from conans.util.env_reader import get_env
 from conans.errors import ConanInvalidConfiguration
 import os
-import re
 import shutil
 import tempfile
 
@@ -34,6 +33,9 @@ class TkConan(ConanFile):
     _source_tcl_subfolder = "sources_tcl"
 
     def configure(self):
+        if self.settings.compiler != "Visual Studio":
+            del self.settings.compiler.libcxx
+
         if self.version.split(".")[:3] != self._tcl_version.split(".")[:3]:
             raise ConanInvalidConfiguration("Versions of tcl and tk do not match")
 
@@ -45,15 +47,8 @@ class TkConan(ConanFile):
         return self.settings.os == "Windows" and self.settings.compiler == "gcc"
 
     def config_options(self):
-        if self.settings.os == "Windows":
+        if self.settings.os == "Windows" or self.options.shared:
             del self.options.fPIC
-        else:
-            if self.options.shared:
-                del self.options.fPIC  # Does not make sense.
-
-    def configure(self):
-        if self.settings.compiler != "Visual Studio":
-            del self.settings.compiler.libcxx
 
     def build_requirements(self):
         if self._is_mingw_windows:
@@ -83,26 +78,28 @@ class TkConan(ConanFile):
 
             os.rename(extracted_dir, source_subfolder)
 
-            unix_config_dir = self._get_configure_dir("unix", source_subfolder)
-            # When disabling 64-bit support (in 32-bit), this test must be 0 in order to use "long long" for 64-bit ints
-            # (${tcl_type_64bit} can be either "__int64" or "long long")
-            tools.replace_in_file(os.path.join(unix_config_dir, "configure"),
-                                  "(sizeof(${tcl_type_64bit})==sizeof(long))",
-                                  "(sizeof(${tcl_type_64bit})!=sizeof(long))")
+            for build_system in ("unix", "win", ):
+                config_dir = self._get_configure_dir(build_system, source_subfolder)
 
-            unix_makefile_in = os.path.join(unix_config_dir, "Makefile.in")
-            # Avoid clearing CFLAGS and LDFLAGS in the makefile
-            tools.replace_in_file(unix_makefile_in, "\nCFLAGS\t", "\n#CFLAGS\t")
-            tools.replace_in_file(unix_makefile_in, "\nLDFLAGS\t", "\n#LDFLAGS\t")
-            tools.replace_in_file(unix_makefile_in, "${CFLAGS}", "${CFLAGS} ${CPPFLAGS}")
+                if build_system != "win":
+                    # When disabling 64-bit support (in 32-bit), this test must be 0 in order to use "long long" for 64-bit ints
+                    # (${tcl_type_64bit} can be either "__int64" or "long long")
+                    tools.replace_in_file(os.path.join(config_dir, "configure"),
+                                          "(sizeof(${tcl_type_64bit})==sizeof(long))",
+                                          "(sizeof(${tcl_type_64bit})!=sizeof(long))")
+
+                makefile_in = os.path.join(config_dir, "Makefile.in")
+                # Avoid clearing CFLAGS and LDFLAGS in the makefile
+                tools.replace_in_file(makefile_in, "\nCFLAGS{}".format(" " if (build_system == "win" and name == "tcl") else "\t"), "\n#CFLAGS\t")
+                tools.replace_in_file(makefile_in, "\nLDFLAGS\t", "\n#LDFLAGS\t")
+                tools.replace_in_file(makefile_in, "${CFLAGS}", "${CFLAGS} ${CPPFLAGS}")
 
         download_tcltk_source(name="tk", filename=filename_tk, url=url_tk, sha256=sha256_tk, extracted_dir="tk{}".format(tk_filename_version), source_subfolder=self._source_subfolder)
         # Building tk on windows, using the tk toolchain requires the tcl sources
         download_tcltk_source(name="tcl", filename=filename_tcl, url=url_tcl, sha256=sha256_tcl, extracted_dir="tcl{}".format(self._tcl_version), source_subfolder=self._source_tcl_subfolder)
 
-    def config_options(self):
-        if self.settings.compiler == "Visual Studio" or self.options.shared:
-            del self.options.fPIC
+        win_makefile_in = os.path.join(self._get_configure_dir("win", self._source_subfolder), "Makefile.in")
+        tools.replace_in_file(win_makefile_in, "\nTCL_GENERIC_DIR", "\n#TCL_GENERIC_DIR")
 
     def system_requirements(self):
         if tools.os_info.with_apt:
@@ -176,17 +173,8 @@ class TkConan(ConanFile):
             )
 
     def _build_autotools(self):
-        # FIXME: move this fixing of tclConfig.sh to tcl
         tcl_root = self.deps_cpp_info["tcl"].rootpath
         tclConfigShPath = os.path.join(tcl_root, "lib", "tclConfig.sh")
-        tools.replace_in_file(tclConfigShPath,
-                              os.path.join(self.package_folder),
-                              tcl_root,
-                              strict=False)
-        tools.replace_in_file(tclConfigShPath,
-                              "TCL_BUILD_",
-                              "#TCL_BUILD_",
-                              strict=False)
 
         conf_args = [
             "--with-tcl={}".format(os.path.dirname(tclConfigShPath.replace("\\", "/"))),
@@ -195,12 +183,14 @@ class TkConan(ConanFile):
             "--enable-symbols" if self.settings.build_type == "Debug" else "--disable-symbols",
             "--enable-64bit" if self.settings.arch == "x86_64" else "--disable-64bit",
             "--with-x" if self.settings.os == "Linux" else "--without-x",
-            "--enable-aqua={}".format("yes" if self.settings.os == "Macos" else "no"),  # autotools fails using aqua
+            "--enable-aqua={}".format("yes" if self.settings.os == "Macos" else "no"),
         ]
 
         autoTools = AutoToolsBuildEnvironment(self, win_bash=tools.os_info.is_windows)
+        if self._is_mingw_windows:
+            autoTools.defines.extend(["UNICODE", "_UNICODE", "_ATL_XP_TARGETING", ])
         autoTools.configure(configure_dir=self._get_configure_dir(), args=conf_args)
-        autoTools.make()
+        autoTools.make(args=["TCL_GENERIC_DIR={}".format(os.path.join(tcl_root, "include")).replace("\\", "/")])
 
     def build(self):
         if self.settings.compiler == "Visual Studio":
@@ -215,12 +205,15 @@ class TkConan(ConanFile):
             with tools.chdir(self.build_folder):
                 autoTools = AutoToolsBuildEnvironment(self, win_bash=tools.os_info.is_windows)
                 autoTools.install()
-            shutil.rmtree(os.path.join(self.package_folder, "lib", "pkgconfig"))
+                autoTools.make(target="install-private-headers")
+            if not self._is_mingw_windows:
+                shutil.rmtree(os.path.join(self.package_folder, "lib", "pkgconfig"))
         self.copy(pattern="license.terms", dst="licenses", src=self._source_subfolder)
         
         tkConfigShPath = os.path.join(self.package_folder, "lib", "tkConfig.sh")
+        pkg_path = os.path.join(self.package_folder).replace('\\', '/')
         tools.replace_in_file(tkConfigShPath,
-                              os.path.join(self.package_folder),
+                              pkg_path,
                               "${TK_ROOT}")
         tools.replace_in_file(tkConfigShPath,
                               "\nTK_BUILD_",
@@ -246,6 +239,10 @@ class TkConan(ConanFile):
             self.cpp_info.exelinkflags.append("-framework Carbon")
             self.cpp_info.exelinkflags.append("-framework IOKit")
             self.cpp_info.sharedlinkflags = self.cpp_info.exelinkflags
+        elif self.settings.os == "Windows":
+            self.cpp_info.libs.extend(["netapi32", "kernel32", "user32", "advapi32", "userenv",
+                                       "ws2_32", "gdi32", "comdlg32", "imm32", "comctl32",
+                                       "shell32", "uuid", "ole32", "oleaut32"])
         
         tk_library = os.path.join(self.package_folder, "lib", "{}{}".format(self.name, ".".join(self.version.split(".")[:2])))
         self.output.info("Setting TCL_LIBRARY environment variable to {}".format(tk_library))
